% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindergarten_readiness.R
\name{get_k_readiness}
\alias{get_k_readiness}
\title{Fetch OSPI Kindergarten readiness scores}
\usage{
get_k_readiness(schoolyears)
}
\arguments{
\item{schoolyears}{vector of initial years of schoolyears, i.e. `c(2023, 2024, 2025)` for multiple years}
}
\value{
data table with combined results from all requested school years
}
\description{
Fetch OSPI Kindergarten readiness scores
}
