% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\name{get_wadata_ospi_url}
\alias{get_wadata_ospi_url}
\title{Identify kindergarten readiness URL}
\usage{
get_wadata_ospi_url(schoolyear)
}
\arguments{
\item{schoolyear}{integer schoolyears}
}
\value{
kindergarten readiness URL
}
\description{
Identify kindergarten readiness URL
}
